% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF2.R
\name{PF2}
\alias{PF2}
\title{PF2}
\usage{
PF2(Q, eps = 1e-10, max_iter = 10000, detail = FALSE)
}
\arguments{
\item{Q}{a square matrix}

\item{eps}{tolerance}

\item{max_iter}{integer, maximum iteration}

\item{detail}{logical, return singular vectors if TRUE}
}
\value{
the largest singular value
}
\description{
Finding the largest singular value of a positive square matrix
}
\examples{
Q = matrix(runif(10*10), 10, 10)
PF2(Q)
}
