% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SK2.R
\name{SK2}
\alias{SK2}
\title{SK2}
\usage{
SK2(A, eps = 1e-10, max_iter = 10000)
}
\arguments{
\item{A}{a square matrix}

\item{eps}{tolerance}

\item{max_iter}{maximum iteration}
}
\value{
a matrix with uniform marginals
}
\description{
Sinkhorn-Knopp algorithm for square matrix
}
\examples{
A = matrix(runif(10*10), 10, 10)
SK2(A)
}
