% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VScor.main.R
\name{VScor.main}
\alias{VScor.main}
\title{VScor.main}
\usage{
VScor.main(A, ...)
}
\arguments{
\item{A}{a square matrix}

\item{...}{options for internal functions}
}
\value{
the singular value of square root of scaled matrix
}
\description{
main function for Variance-Stabilizable association
}
\examples{
A = matrix(runif(10*10), 10, 10)
VScor.main(A)
}
