% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VScor.R
\name{VScor}
\alias{VScor}
\title{VScor}
\usage{
VScor(object, ...)
}
\arguments{
\item{object}{An output of glm}

\item{...}{options for internal functions}
}
\value{
output of specific functions
}
\description{
a wrapper of Variance-Stabilizable correlation
}
\examples{
x = as.table(matrix(c(289, 23, 7, 37), 2, 2))
VScor(x)

glm_1 = glm(dist ~ speed, data=cars)
VScor(glm_1)
}
