% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VScor.table.R
\name{VScor.table}
\alias{VScor.table}
\title{VScor.table}
\usage{
VScor.table(x, ..., draw = FALSE, conf.level = 0.95)
}
\arguments{
\item{x}{a table data}

\item{...}{options for internal functions}

\item{draw}{logical, to draw a plot}

\item{conf.level}{numeric, the confidence level (used only when draw = TRUE)}
}
\value{
list of Riemannian distance (with standard error), variance-stabilizable correlation and squared correlation.
If the size of the table is 2 by 2, the signed Riemannian distance is also returned.
If the option draw is TRUE, confidence intervals are plotted.
}
\description{
Variance-Stabilizable correlation for table data
}
\examples{
N = matrix(c(rpois(2,10), rpois(2,5), rpois(2,10)), 2, 3, byrow=TRUE)
VScor.table(N)
}
