% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VScor.glm.R
\name{VScor.glm}
\alias{VScor.glm}
\title{VScor.glm}
\usage{
VScor.glm(object, ...)
}
\arguments{
\item{object}{An output of glm}

\item{...}{options for internal functions}
}
\value{
a list of association (with standard error), correlation and squared correlation
}
\description{
Variance-Stabilizable association for glm outputs
}
\examples{
glm_1 = glm(dist ~ speed, data=cars)
VScor.glm(glm_1)
}
