% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VScor.R
\name{VScor}
\alias{VScor}
\title{VScor}
\usage{
VScor(object, ...)
}
\arguments{
\item{object}{An output of glm}

\item{...}{options for internal functions}
}
\value{
output of specific functions
}
\description{
The package VScor computes the variance-stabilizable (VS) correlation coefficient of table data or output of glm.
The method is based on the Riemannian distance from the independence model.
The Riemannian distance has the standard error 1/sqrt(n), where n is the sample size.
The VS correlation is a transformation of the Riemannian distance that coincides with the Pearson correlation coefficient when the population is bivariate Gaussian.
If the input is a 2 by 2 table, VScor also returns a signed Riemannian distance.

The function VScor() is a wrapper of VScor.glm and VScor.table
}
\examples{
x = as.table(matrix(c(289, 23, 7, 37), 2, 2))
VScor(x)

glm_1 = glm(dist ~ speed, data=cars)
VScor(glm_1)
}
