% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VScor.table.R
\name{VScor.table}
\alias{VScor.table}
\title{VScor.table}
\usage{
VScor.table(x, ...)
}
\arguments{
\item{x}{a table data}

\item{...}{options for internal functions}
}
\value{
list of variance-stabilized association (with standard error), correlation and squared correlation
}
\description{
Variance-Stabilizable association for table data
}
\examples{
N = matrix(c(rpois(2,10), rpois(2,5), rpois(2,10)), 2, 3, byrow=TRUE)
VScor.table(N)
}
